chore: move reviewable pull requests section to its folder
