+++ tests/views/Guarded/Dashboard/DashboardView.spec.ts
@@ -1,8 +1,8 @@
 import DashboardView from '@/views/Guarded/Dashboard/DashboardView.vue'
 import RecentCommitsWithoutPullRequest from '@/views/Guarded/Dashboard/RecentCommitsWithoutPullRequest.vue'
-import ReviewablePullRequests from '@/views/Guarded/Dashboard/ReviewablePullRequests.vue'
-import { shallowMount } from '@vue/test-utils'
+import ReviewablePullRequests from '@/views/Guarded/Dashboard/ReviewablePullRequests/ReviewablePullRequests.vue'
 import { createTestingPinia } from '@pinia/testing'
+import { shallowMount } from '@vue/test-utils'
 
 describe.concurrent('Dashboard View', () => {
 	const wrapper = shallowMount(DashboardView, {
