+++ tests/views/Guarded/Dashboard/ReviewablePullRequests.spec.ts
@@ -1,8 +1,8 @@
 import { useNavigationStore } from '@/stores/NavigationStore'
 import { usePullRequestStore } from '@/stores/PullRequestStore'
 import { useRepositoryStore } from '@/stores/RepositoryStore'
-import ReviewablePullRequests from '@/views/Guarded/Dashboard/ReviewablePullRequests.vue'
 import PullRequestRow from '@/views/Guarded/Dashboard/ReviewablePullRequests/PullRequestRow.vue'
+import ReviewablePullRequests from '@/views/Guarded/Dashboard/ReviewablePullRequests/ReviewablePullRequests.vue'
 import { createTestingPinia } from '@pinia/testing'
 import PullRequestFixture from '@tests/__fixtures__/PullRequestFixture.json'
 import RepositoryFixture from '@tests/__fixtures__/RepositoryFixture.json'
